+++ bapsflib/_hdf/maps/controls/tests/__init__.py
@@ -8,6 +8,5 @@
 # License: Standard 3-clause BSD; see "LICENSES/LICENSE.txt" for full
 #   license terms and contributor agreement.
 #
-from bapsflib._hdf.maps.controls.tests.fauxpositions180e import FauxPositions180E
 from bapsflib._hdf.maps.controls.tests.fauxsixk import FauxSixK
 from bapsflib._hdf.maps.controls.tests.fauxwaveform import FauxWaveform
