+++ bapsflib/_hdf/maps/controls/tests/test_positions180e.py
@@ -3,10 +3,9 @@ import numpy as np
 
 from bapsflib._hdf.maps.controls.positions180e import HDFMapControlPositions180E
 from bapsflib._hdf.maps.controls.tests.common import ControlTestCase
-from bapsflib._hdf.maps.controls.tests.fauxpositions180e import FauxPositions180E
+from bapsflib.faux.controls.fauxpositions180e import FauxPositions180E
 from bapsflib._hdf.maps.controls.types import ConType
 from bapsflib.utils.exceptions import HDFMappingError
-from bapsflib.utils.warnings import HDFMappingWarning
 
 
 class TestPositions180E(ControlTestCase):
