+++ bapsflib/faux/controls/fauxpositions180e.py
